+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/spell/Spell.java
@@ -123,4 +123,5 @@ public interface Spell extends SpellTemplate {
     boolean isEnabled();
     void setEnabled(boolean enabled);
     long getLastCast();
+    double getChargesRemaining();
 }